[chore] pagination dto compact
